pyproject.toml
[project]
name = "gruner-social-features"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "huggingface-hub",
  "diffusers",
  "accelerate",
  "torch",
  "pillow",
  "python-dotenv",
  "google-genai==1.68.0",
  "transformers>=5.3.0",
  "boto3>=1.42.76",
  "openai>=2.30.0",
]
